:lipstick: Add kitty terminal compatibility
The terminal kitty handles emoji widths different to most other
terminals. Any emojis with the variant selector needs special handling.

Emojis that are single width but display as double width and are marked
as variants need to be displayed as two visible characters. This
involves considering them as one character wider.

This however does not apply to emojis marked as variants that are
single width and displayed as a single character.